я сделала оптимизацию
