+++ gulpfile.js
@@ -2,28 +2,110 @@ import gulp from 'gulp';
 import plumber from 'gulp-plumber';
 import sass from 'gulp-dart-sass';
 import postcss from 'gulp-postcss';
+import csso from 'postcss-csso';
 import autoprefixer from 'autoprefixer';
 import browser from 'browser-sync';
+import rename from 'gulp-rename';
+import terser from 'gulp-terser';
+import squoosh from 'gulp-libsquoosh';
+import svgo from 'gulp-svgmin';
+import svgstore from 'gulp-svgstore';
+import { deleteSync } from 'del';
+import htmlmin from 'gulp-htmlmin';
 
 // Styles
 
-export const styles = () => {
-  return gulp.src('source/sass/style.scss', { sourcemaps: true })
+export const styles = (done) => {
+  gulp.src('source/sass/style.scss', { sourcemaps: true })
     .pipe(plumber())
     .pipe(sass().on('error', sass.logError))
     .pipe(postcss([
-      autoprefixer()
+      autoprefixer(),
+      csso()
     ]))
-    .pipe(gulp.dest('source/css', { sourcemaps: '.' }))
+    .pipe(rename('style.min.css'))
+    .pipe(gulp.dest('build/css', { sourcemaps: '.' }))
     .pipe(browser.stream());
+    done()
+}
+
+export const html = (done) => {
+   gulp.src('source/*.html')
+   .pipe(htmlmin({
+    removeComments: true,
+    collapseWhitespace: true
+  }))
+  .pipe(gulp.dest('build'));
+  done()
+}
+
+const script = (done ) => {
+   gulp.src('source/js/*.js')
+  .pipe(terser())
+  .pipe(gulp.dest('build/js'));
+  done()
+}
+
+const svg = (done) =>{
+gulp.src(['source/img/*.svg', '!source/img/icons/*.svg'])
+.pipe(svgo())
+.pipe(gulp.dest('build/img'));
+done()
+}
+const sprite =(done) =>{
+gulp.src('source/img/icons/*.svg')
+.pipe(svgo())
+.pipe(svgstore({
+  inlineSvg:true
+}))
+.pipe(rename('sprite.svg'))
+.pipe(gulp.dest('build/img'));
+done()
+}
+
+const optimzeImages = () => {
+  return gulp.src('source/img/**/*.{jpg,png}')
+    .pipe(gulp.dest('build/img'));
+}
+
+const copyImages =(done) => {
+  gulp.src('source/img/**/*.{jpg,png,webp}')
+  .pipe(gulp.dest('build/img'));
+  done()
+}
+
+export const createWebp =(done) => {
+  gulp.src('source/img/**/*.{jpg,png}')
+  .pipe(squoosh({
+    webp: {}
+  }))
+  .pipe(gulp.dest('build/img'));
+  done()
+}
+
+const copy = done => {
+  gulp.src([
+    'source/fonts/*.{woff2,woff}',
+    'source/*.ico',
+    'source/*.webmanifest'
+  ], {
+    base:'source'
+  })
+  .pipe(gulp.dest('build'));
+  done()
+};
+
+const clean = (done) =>{
+  deleteSync('build');
+  done()
 }
 
 // Server
 
-const server = (done) => {
+const server = async (done) => {
   browser.init({
     server: {
-      baseDir: 'source'
+      baseDir: 'build'
     },
     cors: true,
     notify: false,
@@ -34,12 +116,41 @@ const server = (done) => {
 
 // Watcher
 
-const watcher = () => {
+const watcher = async () => {
   gulp.watch('source/sass/**/*.scss', gulp.series(styles));
-  gulp.watch('source/*.html').on('change', browser.reload);
+  gulp.watch('source/js/**/*.js', gulp.series(script));
+  gulp.watch('source/*.html', gulp.series(html));
 }
 
 
-export default gulp.series(
-  styles, server, watcher
+export const build = gulp.series(
+  clean,
+  copy,
+  copyImages,
+  gulp.parallel(
+    styles,
+    html,
+    script,
+    svg,
+    sprite,
+    createWebp
+  ),
 );
+
+export default gulp.series(
+  clean,
+  copy,
+  copyImages,
+  optimzeImages,
+  gulp.parallel(
+    styles,
+    html,
+    script,
+    svg,
+    sprite,
+    createWebp
+  ),
+  server,
+  watcher
+);
+
