+++ source/form.html
@@ -3,13 +3,15 @@
   <head>
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <link rel="stylesheet" href="./css/style.css" />
-    <!-- <link rel="stylesheet" href="./css/style.min.css" /> -->
+    <!-- <link rel="stylesheet" href="./css/style.css" /> -->
+    <link rel="stylesheet" href="./css/style.min.css" />
     <link rel="icon" href="favicon.ico" />
     <link rel="icon" href="./img/favicons/favicon.svg" type="image/svg+xml" />
     <link rel="apple-touch-icon" href="./img/favicons/apple.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
+    <link rel="preload" href="./fonts/ptsans.woff2" as="font" type="font/woff2" crossorigin>
+    <link rel="preload" href="./fonts/ptsansbold.woff2" as="font" type="font/woff2" crossorigin>
   </head>
   <body>
     <svg style="display: none">
