+++ source/index.html
@@ -4,11 +4,13 @@
     <meta charset="utf-8" />
     <title>HTML Academy: Седона</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="stylesheet" href="./css/style.css" />
-    <!-- <link rel="stylesheet" href="./css/style.min.css" /> -->
+    <!-- <link rel="stylesheet" href="./css/style.css" /> -->
+    <link rel="stylesheet" href="./css/style.min.css" />
     <link rel="icon" href="favicon.ico" />
     <link rel="icon" href="./img/favicons/favicon.svg" type="image/svg+xml" />
     <link rel="apple-touch-icon" href="./img/favicons/apple.png" />
+    <link rel="preload" href="./fonts/ptsans.woff2" as="font" type="font/woff2" crossorigin>
+    <link rel="preload" href="./fonts/ptsansbold.woff2" as="font" type="font/woff2" crossorigin>
   </head>
   <body>
     <svg style="display: none">
